pdf_search_tool/requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
PyMuPDF
opencv-python
langdetect